添加frozen,更新API
